고객센터 질의 및 응답 구현방식 변경
- 불필요한 필드 삭제: isAnswered
- 응답 삭제 로직에서 Question의 연관관계 해제를 위해 questionId를 받도록 수정

+++ module-api/src/main/java/com/junior/service/qna/AnswerService.java
@@ -85,10 +85,21 @@ public class AnswerService {
             throw new NotValidMemberException(StatusCode.INVALID_MEMBER_STATUS);
         }
 
-        Answer answer = answerRepository.findById(questionId)
-                .orElseThrow(() -> new AnswerException(StatusCode.ANSWER_NOT_FOUND));
+        Question question = questionRepository.findById(questionId)
+                .orElseThrow(() -> new QuestionException(StatusCode.QUESTION_NOT_FOUND));
+
+        Answer answer = question.getAnswer();
+
+        if (answer == null) {
+            throw new AnswerException(StatusCode.ANSWER_NOT_FOUND);
+        }
+
+
 
         answer.delete();
+
+        //연관관계 해제
+        question.deleteAnswer();
 
 
     }
